+++ components/Navbar.tsx
@@ -4,6 +4,8 @@ import Link from "next/link";
 import styles from "../styles/components/Navbar.module.scss";
 import { useCartStore } from "@/store/cart";
 import { Session } from "next-auth";
+import { useAuthStore } from "@/store/authStore";
+import { useEffect } from "react";
 
 type Props = {
 	session: Session | null;
@@ -11,8 +13,21 @@ type Props = {
 
 export default function Navbar({ session }: Props) {
 	const itemCount = useCartStore((state) => state.items.reduce((sum, item) => sum + item.quantity, 0));
-	const isLoggedIn = !!session?.user;
-	console.log("session:", session);
+	const setAuthDetails = useAuthStore((state) => state.setAuthDetails);
+	const setIsLoggedIn = useAuthStore((state) => state.setIsLoggedIn);
+	const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
+	const role = useAuthStore((state) => state.role);
+	const isAdmin = role === "ADMIN";
+
+	useEffect(() => {
+		if (!!session?.user) {
+			setAuthDetails(session.user.email, session.user.role);
+		} else {
+			setIsLoggedIn(false);
+			setAuthDetails(null, "USER");
+		}
+		console.log(session)
+	}, [session]);
 
 	return (
 		<nav className={styles.navbar}>
@@ -21,29 +36,39 @@ export default function Navbar({ session }: Props) {
 			</div>
 			<ul className={styles.links}>
 				<li>
-					<Link href="/">Home</Link>
+					<Link href="/">Anasayfa</Link>
 				</li>
 				<li>
-					<Link href="/menu">Menu</Link>
+					<Link href="/menu">Menü</Link>
 				</li>
-				<li>
-					<Link href="/about">About</Link>
-				</li>
-				<li>
-					<Link href="/cart">Cart{itemCount > 0 && <span className={styles.cartCount}>{itemCount}</span>}</Link>
-				</li>
+				{!isAdmin && (
+					<>
+						<li>
+							<Link href="/about">Hakkında</Link>
+						</li>
+
+						<li>
+							<Link href="/cart">Sepet{itemCount > 0 && <span className={styles.cartCount}>{itemCount}</span>}</Link>
+						</li>
+					</>
+				)}
+				{isAdmin && (
+					<li>
+						<Link href="/admin/products">Ürün Yönetimi</Link>
+					</li>
+				)}
 
 				{isLoggedIn ? (
 					<li>
-						<Link href="/profile">Profile</Link>
+						<Link href="/profile">Hesap</Link>
 					</li>
 				) : (
 					<>
 						<li>
-							<Link href="/signin">Sign In</Link>
+							<Link href="/signin">Giriş Yap</Link>
 						</li>
 						<li>
-							<Link href="/signup">Sign Up</Link>
+							<Link href="/signup">Kayıt Ol</Link>
 						</li>
 					</>
 				)}
