+++ styles/globals.scss
@@ -6,7 +6,8 @@ body {
 	padding: 0;
 	// background: #fafafa;
 	//background-image: url("https://www.tilda.com/wp-content/uploads/2024/07/shutterstock_562247284-scaled.jpeg");
-	background-image: linear-gradient(rgba(0, 0, 0, 0.527), rgba(0, 0, 0, 0.5)),
+	background-image:
+		linear-gradient(rgba(0, 0, 0, 0.527), rgba(0, 0, 0, 0.5)),
 		url("https://www.tilda.com/wp-content/uploads/2024/07/shutterstock_562247284-scaled.jpeg");
 	background-repeat: no-repeat;
 	background-attachment: fixed;
@@ -23,6 +24,10 @@ h4,
 h5,
 h6 {
 	font-weight: 600;
+}
+
+button:disabled {
+	opacity: 0.3;
 }
 
 @keyframes skeleton-loading {
